pyproject.toml
[project]
name = "rag-llm-applications"
version = "2.0"
description = "An illustration of LLMs and RAG."
readme = "README.md"
requires-python = ">3.10, <3.12"
authors = [
    { name = "Tore Erdmann", email = "tore.erdmann@alexanderthamm.com" },
    { name = "Philipp Schwartenbeck", email = "philipp.schwartenbeck@alexanderthamm.com" },
]
license = { text = "Proprietary" }
dependencies = [
    "PyYAML>=6.0,<7.0",
    "langchain",
    "qdrant-client>=1.5.4,<2.0.0",
    "pypdf>=3.16.2,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "tiktoken",
    "streamlit>=1.27.2,<2.0.0",
    "transformers>=4.34.0,<5.0.0",
    "pymupdf>=1.23.6,<2.0.0",
    "seaborn>=0.13.0,<0.14.0",
    "scikit-learn>=1.3.2,<1.4.0",
    "nltk>=3.8.1,<3.9.0",
    "rouge>=1.0.1,<1.1.0",
    "sentence-transformers>=2.2.2,<2.3.0",
    "langchain-openai",
    "deepeval>=2.7.6",
    "ipykernel>=6.29.5",
    "docling-core>=2.28.1",
    "docling>=2.31.0",
    "langchain-community>=0.3.26",
    "ipywidgets>=8.1.7",
]

[dependency-groups]
dev = [
    "black>=22.10,<23.0",
    "pre-commit>=2.20,<3.0",
    "pytest>=7.2,<8.0",
    "pytest-cov>=4.0,<5.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
ignore = ["F401"]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["rag_llm_applications"]

[tool.setuptools.package-data]
rag_llm_applications = ["res/*"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
